Fix DVFS operating points for 144/312 MHz
Per the [A20 SDK] (https://linux-sunxi.org/A20) (see section DVFS) the DVFS OPP voltages for 144000 should be 1050000mV and for 312000 it should be 1100000mV. This commit changes the voltages to the values recommended. There was a [commit already in 2015](https://github.com/torvalds/linux/commit/a15b80fb0ad7ddab8d2668392cce3c12aff594bd) which changed voltages accordingly - however only specifically for the bananapi. 

Originally OPP voltages were set to 1000000mV in this file. The consequence of running the CPU at 1000000mV for 144000 and 312000 respectively is instable operation and regular crashes.